@@ -22,100 +22,107 @@
   },
   "private": true,
   "dependencies": {
-    "@agm/core": "^1.0.0",
-    "@angular/animations": "7.0.2",
-    "@angular/cdk": "7.0.2",
-    "@angular/common": "7.0.2",
-    "@angular/compiler": "7.0.2",
-    "@angular/core": "7.0.2",
-    "@angular/forms": "7.0.2",
-    "@angular/http": "7.0.2",
-    "@angular/material": "7.0.2",
-    "@angular/platform-browser": "7.0.2",
-    "@angular/platform-browser-dynamic": "7.0.2",
-    "@angular/platform-server": "7.0.2",
-    "@angular/router": "7.0.2",
-    "@ngrx/effects": "^7.0.1",
-    "@ngrx/entity": "^7.0.1",
-    "@ngrx/router-store": "^7.0.1",
-    "@ngrx/store": "^7.0.1",
-    "@ngrx/store-devtools": "^7.0.1",
-    "@types/base-64": "^0.1.3",
+    "@agm/core": "^1.1.0",
+    "@angular/animations": "8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "8.2.14",
+    "@angular/compiler": "8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "8.2.14",
+    "@angular/http": "^7.2.15",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "8.2.14",
+    "@angular/platform-server": "8.2.14",
+    "@angular/router": "8.2.14",
+    "@ngrx/effects": "^8.6.0",
+    "@ngrx/entity": "^8.6.0",
+    "@ngrx/router-store": "^8.6.0",
+    "@ngrx/store": "^8.6.0",
+    "@ngrx/store-devtools": "^8.6.0",
+    "@types/bignumber.js": "^5.0.0",
+    "@types/bluebird": "^3.5.29",
     "@types/jquery": "^3.3.31",
     "@types/utf8": "^2.1.6",
-    "ajv": "6.4.0",
+    "ajv": "^6.10.2",
     "arrive": "2.4.1",
-    "base-64": "^0.1.0",
-    "bip32": "git+https://github.com/bitcoinjs/bip32.git#patchRmd160-2",
+    "bip32": "^2.0.4",
     "bluebird": "^3.7.2",
-    "bootstrap": "4.3.1",
-    "bootstrap-material-design": "4.1.1",
-    "bootstrap-notify": "3.1.3",
-    "chartist": "0.11.0",
-    "classlist.js": "1.1.20150312",
-    "core-js": "2.4.1",
-    "ethereumjs-tx": "^1.3.7",
-    "express": "4.16.3",
+    "bootstrap-material-design": "^4.1.2",
+    "bootstrap-notify": "^3.1.3",
+    "braces": "^3.0.2",
+    "chartist": "^0.11.4",
+    "classlist.js": "^1.1.20150312",
+    "core-js": "^3.6.1",
+    "ethereumjs-tx": "^2.1.2",
+    "express": "^4.17.1",
     "flag-icon-css": "^3.4.5",
-    "fusioncharts": "^3.15.0",
+    "fusioncharts": "^3.15.0-sr.1",
     "fusionmaps": "^3.15.0",
-    "googleapis": "28.1.0",
-    "hammerjs": "2.0.8",
+    "googleapis": "^46.0.0",
+    "hammerjs": "^2.0.8",
     "intl": "^1.2.5",
     "jquery": "^3.4.1",
-    "moment": "2.22.1",
+    "mini-css-extract-plugin": "^0.9.0",
+    "moment": "^2.24.0",
     "ngx-uploadcare-widget": "^1.2.5",
-    "node-pre-gyp": "^0.13.0",
-    "npm-packlist": "^1.4.6",
-    "nyc": "^14.1.1",
+    "node-pre-gyp": "^0.14.0",
+    "node-sass": "^4.13.0",
+    "npm-packlist": "^2.0.3",
+    "nyc": "^15.0.0",
     "perfect-scrollbar": "^1.4.0",
-    "popper.js": "1.14.3",
-    "react-images-uploader": "^1.2.0-rc1",
-    "rxjs": "6.3.3",
-    "rxjs-compat": "6.3.3",
-    "truffle-contract": "^4.0.4",
+    "popper.js": "^1.16.0",
+    "postcss": "^7.0.25",
+    "postcss-import": "^12.0.1",
+    "postcss-load-config": "^2.1.0",
+    "postcss-loader": "^3.0.0",
+    "postcss-value-parser": "^4.0.2",
+    "rxjs": "^6.5.3",
+    "rxjs-compat": "^6.5.3",
+    "secp256k1": "^3.8.0",
+    "truffle-contract": "^4.0.31",
     "ts-type": "^1.2.9",
     "utf8": "^3.0.0",
-    "web-animations-js": "2.3.1",
+    "web-animations-js": "^2.3.2",
     "web3": "^1.2.4",
-    "zone.js": "0.8.26"
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
     "@angular-builders/custom-webpack": "^8.4.1",
     "@angular-builders/dev-server": "^7.3.1",
     "@angular-devkit/build-angular": "^0.803.21",
+    "@angular-devkit/architect": "^0.803.21",
     "@angular/cli": "^8.3.21",
     "@angular/compiler-cli": "^8.2.14",
-    "@angular/language-service": "7.0.2",
-    "@types/bootstrap": "3.3.32",
-    "@types/chartist": "0.9.34",
-    "@types/googlemaps": "3.30.8",
-    "@types/jasmine": "2.5.38",
-    "@types/node": "^6.14.9",
+    "@angular/language-service": "8.2.14",
+    "@types/bootstrap": "^4.3.1",
+    "@types/chartist": "^0.9.46",
+    "@types/googlemaps": "^3.39.0",
+    "@types/jasmine": "^3.5.0",
+    "@types/node": "^13.1.0",
     "babel-cli": "^6.26.0",
     "babel-preset-es2015": "^6.24.1",
-    "bootstrap": "^4.1.0",
-    "chai": "^4.1.2",
+    "bootstrap": "^4.4.1",
+    "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
     "chai-bignumber": "^3.0.0",
-    "codelyzer": "4.2.1",
+    "codelyzer": "^5.2.1",
     "dotenv": "^8.2.0",
-    "ganache-cli": "^6.1.8",
-    "jasmine-core": "3.1.0",
-    "jasmine-spec-reporter": "4.2.1",
-    "karma": "2.0.0",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-cli": "1.0.1",
-    "karma-coverage-istanbul-reporter": "1.4.2",
-    "karma-jasmine": "1.1.1",
-    "lite-server": "^2.3.0",
-    "mini-css-extract-plugin": "^0.9.0",
-    "nodemon": "^1.19.4",
-    "protractor": "^6.0.0",
+    "ganache-cli": "^6.7.0",
+    "jasmine-core": "^3.5.0",
+    "jasmine-spec-reporter": "^4.2.1",
+    "karma": "^4.4.1",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-cli": "^2.0.0",
+    "karma-coverage-istanbul-reporter": "^2.1.1",
+    "karma-jasmine": "^2.0.1",
+    "lite-server": "^2.5.4",
+    "nodemon": "^2.0.2",
+    "protractor": "^5.4.2",
     "solidity-coverage": "^0.6.7",
-    "ts-node": "5.0.1",
-    "tslint": "5.9.1",
-    "typescript": "3.1.6",
+    "ts-node": "^8.5.4",
+    "tslint": "^5.20.1",
+    "typescript": "3.5.3",
     "webpack-merge": "^4.2.2"
   }
 }
