@@ -34,14 +34,14 @@ export class DistributorserviceService {
       } catch (error) {
         console.log('error');
       }
-        if (this.result != null){
+        if (this.result != null) {
         this.sendPasswordResetLink(this.result.logs[0].args.email, this.result.logs[0].args.passwordRestToken);
           dt(this.result.logs[0].args.name);
       } else {
         er('user alredy in system');
         console.log(this.result)
       }
-    });});
+    }); });
   }
   updateDistributorName(email, name): Promise<any> {
     // tslint:disable-next-line: max-line-length
@@ -112,17 +112,17 @@ export class DistributorserviceService {
         });
     });
   }
-  updateDistributor(email,index, name, adress, telepone): Promise<any> {
+  updateDistributor(email, index, name, adress, telepone): Promise<any> {
     // tslint:disable-next-line: max-line-length
     // tslint:disable-next-line: no-unused-expression
     return new Promise((dt, er) => {
       this.web3.setDistributorvalue()
         .then(async (deployed: {
-          updateDistributor: (arg0: string,arg1: number[], arg2: string, arg3: string, arg4: string, arg5: { from: string; }) =>
+          updateDistributor: (arg0: string, arg1: number[], arg2: string, arg3: string, arg4: string, arg5: { from: string; }) =>
             { er: string; ev: string; }
         }) => {
           try {
-            this.result = await deployed.updateDistributor(email,index, name, adress, telepone ,{ from: this.web3.curentaccount })
+            this.result = await deployed.updateDistributor(email, index, name, adress, telepone , { from: this.web3.curentaccount })
           } catch (error) {
             console.log('error');
           }
@@ -134,7 +134,6 @@ export class DistributorserviceService {
           }
         });
     });
-  
   }
   deleteDistributor(email): Promise<any> {
     // tslint:disable-next-line: max-line-length
@@ -146,7 +145,7 @@ export class DistributorserviceService {
             { er: string; ev: string; }
         }) => {
           try {
-            this.result = await deployed.deleteDistributor(email,{ from: this.web3.curentaccount })
+            this.result = await deployed.deleteDistributor(email, { from: this.web3.curentaccount })
           } catch (error) {
             console.log('error');
           }
@@ -201,8 +200,8 @@ export class DistributorserviceService {
     });
     return this.result;
   }
-  sendPasswordResetLink( _email,_data) {
-    console.log(_data + '   ' +_email);
+  sendPasswordResetLink( _email, _data) {
+    console.log(_data + '   ' + _email);
     return true;
   }
 }
